// src/types/shopMainInfo.tsx
export interface ShopMainInfo {
  shop_name: string;
  logo_path: string;
  description: string;
  cities: number[];
  goods_count: number;
  followers_count: number;
  phone_numbers: string[];
  whatsapp: string;
  addresses: string[];
  categories: number[];
  instagram: string;
  shop_id?: number;
}